Create AppLayoutSidebarNavListItem and stories
